Introduced the option to use PyTorch's spectral gating for noise reduction in the web UI.

+++ functions/filter.py
@@ -1,8 +1,9 @@
 import os
 import noisereduce as nr
-import librosa
 import soundfile as sf
 import numpy as np
+
+from scipy.io import wavfile
 
 from functions.helper.run_san import check_wav_files
 
@@ -54,48 +55,50 @@ class NoiseReducer:
         #return audio and logs
         return reduced_audio, logs
 
-    def process_single_audio_file(self, file, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal):
+    def process_single_audio_file(self, file, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal, use_hardware_acceleration):
         file_path = os.path.join(self.input_dir, file)
-        audio_data, sample_rate = librosa.load(file_path, sr=None)
-
-        yield f"[DEBUG] Processing {file_path}. Sample rate: {sample_rate}, Audio length: {len(audio_data)}"
-
-        #apply reduct and get logs
-        reduced_audio, reduction_logs = self.apply_dynamic_noise_reduction(
-            audio_data, sample_rate, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal)
-
-        #yield log
-        for log in reduction_logs:
-            yield log
-
-        #save processed file as new
         new_filename = file.replace('.wav', '_cleaned.wav')
         new_file_path = os.path.join(self.input_dir, new_filename)
-        sf.write(new_file_path, reduced_audio, sample_rate)
-        os.remove(file_path)
+
+        yield f"[DEBUG] Processing {file_path}."
+
+        sample_rate, audio_data = wavfile.read(file_path)
+        yield f"Sample rate: {sample_rate}, Audio length: {len(audio_data)}"
+        
+        if use_hardware_acceleration:
+            reduced_audio = nr.reduce_noise(y=audio_data, sr=sample_rate, nonstationary=True)
+            wavfile.write(new_file_path, sample_rate, reduced_audio)
+
+        else:
+            reduced_audio, reduction_logs = self.apply_dynamic_noise_reduction(
+                audio_data, sample_rate, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal)
+
+            yield from reduction_logs
+            wavfile.write(new_file_path, sample_rate, reduced_audio)
+            os.remove(file_path)
 
         yield f"[DEBUG] Saved cleaned file as {new_file_path} and removed original file {file_path}.\n=====================================\n"
 
-    def process_audio_files(self, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal):
+    def process_audio_files(self, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal, use_hardware_acceleration):
         files = [f for f in os.listdir(self.input_dir) if f.endswith('.wav') and not f.endswith('_cleaned.wav')]
         yield f"[DEBUG] Found {len(files)} files to process.\n"
 
         # Process files sequentially to ensure proper yielding to Gradio
         for file in files:
             try:
-                for log in self.process_single_audio_file(file, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal):
+                for log in self.process_single_audio_file(file, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal, use_hardware_acceleration=False):
                     yield log
             except Exception as e:
                 yield f"[ERROR] Failed to process {file}: {str(e)}"
 
         yield f"\n[OK] Finished filtering {len(files)} files."
 
-    def gradio_run(self, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal):
+    def gradio_run(self, frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal, use_hardware_acceleration):
         if not check_wav_files():
             yield "ERROR: No .wav files found in the input directory. Please upload them and try again."
             return
 
         logs = []
-        for log in self.process_audio_files(frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal):
+        for log in self.process_audio_files(frame_length, hop_length, silence_threshold, prop_decrease_noisy, prop_decrease_normal, use_hardware_acceleration):
             logs.append(log)
             yield "\n".join(logs)
